requirements.txt
langchain
langchain_openai
python-dotenv
psycopg2-binary